requirements.txt
gunicorn
psycopg2-binary
pytz==2025.2
tzdata==2025.2
docutils
Django==5.*
django-extensions
djangorestframework
djangorestframework-simplejwt
drf-spectacular
django-environ
django-vite
django-admin-interface
Pillow
easy-thumbnails
django-filer
django-admin-sortable2
djangoql
markdown
django-tinymce
django-cotton
